+++ sync-for-expenses/src/sdk/models/operations/updatesupplier.ts
@@ -10,9 +10,21 @@ export class UpdateSupplierRequest extends SpeakeasyBase {
     @SpeakeasyMetadata({ data: "request, media_type=application/json" })
     supplier?: shared.Supplier;
 
+    /**
+     * Allow a sync upon push completion.
+     */
+    @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=allowSyncOnPushComplete" })
+    allowSyncOnPushComplete?: boolean;
+
+    /**
+     * Unique identifier for a company.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=companyId" })
     companyId: string;
 
+    /**
+     * Unique identifier for a connection.
+     */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=connectionId" })
     connectionId: string;
 
@@ -23,16 +35,22 @@ export class UpdateSupplierRequest extends SpeakeasyBase {
     forceUpdate?: boolean;
 
     /**
-     * Unique identifier for a supplier
+     * Unique identifier for a supplier.
      */
     @SpeakeasyMetadata({ data: "pathParam, style=simple;explode=false;name=supplierId" })
     supplierId: string;
 
+    /**
+     * Time limit for the push operation to complete before it is timed out.
+     */
     @SpeakeasyMetadata({ data: "queryParam, style=form;explode=true;name=timeoutInMinutes" })
     timeoutInMinutes?: number;
 }
 
 export class UpdateSupplierResponse extends SpeakeasyBase {
+    /**
+     * HTTP response content type for this operation
+     */
     @SpeakeasyMetadata()
     contentType: string;
 
@@ -42,9 +60,15 @@ export class UpdateSupplierResponse extends SpeakeasyBase {
     @SpeakeasyMetadata()
     errorMessage?: shared.ErrorMessage;
 
+    /**
+     * HTTP response status code for this operation
+     */
     @SpeakeasyMetadata()
     statusCode: number;
 
+    /**
+     * Raw HTTP response; suitable for custom response parsing
+     */
     @SpeakeasyMetadata()
     rawResponse?: AxiosResponse;